requirements.txt
Django~=5.0a1
djangorestframework~=3.14.0
Pillow~=10.1.0
djoser~=2.2.0
drf_spectacular_extensions~=0.0.2
djangorestframework-simplejwt~=5.3.0
python-decouple~=3.8.0
PyYAML~=6.0.1